New Feature: Packing Activity Directives With Offset (#1752)
* Beginning of PackLeftAction

Working towards creating the packLeft function, and helper methods to parse Stings of format "HH:mm:ss.SSSSSS" into microseconds. There seems to be some issues with the hasura query and the format of the duration string that needs to be reworked.

* Pack Left

Have an early working implementation of packLeft, it currently works on the framework of activities can only be packed if they have the same anchor, "null" is a valid anchor.

* Revamping packLeft

Working on a new implementation of packLeft that allows packing of various anchored activities but follows a specific set of defined rules discussed in a meeting. Will add documentation later to explain the rules

* Reworked Pack Left

Have a new working implementation of packLeft that covers the edge cases of anchors. Some refactoring still needs to be done and then replication for packRight, also realized that initialStartTime is only necessary to keep track of for the first activity and not a map to all activities.

* Reset Pack Activities

Revamped packLeft with the new guidelines involving anchor activities discussed in the Aerie-External Sim Shallow Integration Working Group Meeting. This new approach is in a function with the temporary name packActivityDirectivesInPlanRevamp. I have some ideas to use the same method for left and right with an enum switch between the two but that is something I am still testing, hence the reason it is in a temporary named funciton for now.

* fixed packRight and refactored the appropriate packLeft

Have a working implementation of packRight and updated condition check for if the activity being packed was an anchor to other activities (it initally didn't do anything because it was the wrong conditional). Removed the old packLeftActivityDirectivesInPlan implementation and refactored the "revamp" method to the appropriate name.

* Introduced a Less cluttered version of packing Activities

Combined both implementations of packing Activities in a single method using a const union type for direction. Will be talking with Aerie-UI team if this is fine for the development standards, for now not deleting the code for the separate options as well.

* Started building PackActivitiesOffsetSelect

Built a dialog for packing activity directive with offset, the cancel functionality works but the pack button still needs to be updated. First need to make some refactoring changes to pack actions.

* Code cleanup for packActivityDirectivesInPlan

Deleted the separate methods for packing activities to left and right, and kept packActivityDirectivesInPlan which handles both cases. Removed unnecessary code and used aerie-ui stores instead to achieve same functionality.

* Added offset functionality

Instantiated an offset field (in microseconds) to packing Activities. This will play a more vital role in the development of the pack-actions-offset feature that is also being developed. When the regular 'Pack Left' and 'Pack Right' options are selected the offset field is 0.

* Working Implementation of Pack with Offset

Have a working implementation of packing activities with offset. A dialog pops up that allows the user to pick a direction and offset (default to 0).

* Updated design for PackActivitiesOffsetSelect

Changed the offset string box to the appropriate CSS for error, don't allow negative offsets to be entered. The user cannot pack if the Offset string input is invalid.

* Changed location of where database gets updated

Felt it would be more appropriate if the database was updated from ActivityDirectivesTable and packActivityDirectivesInPlan just returns the array of updated activities instead. This also simplifies when writing unit tests

* Update Packing with Offset to match new packing framework

Updated packActivityWithOffset so it the updates to the database are done in the ActivityDirectivesTable, also simplified updating an array of ActivityDirectives into one method

* style: cleaned out code, refactor: design new modal

Design a new modal for pack Activities with offset that utilizes Aerie's Modal to match with design changes. Cleaned out unnecessary code.

* style: cleanup and rename

Change name from packActivityDirectivesBothInPlan to just packActivityDirectivesInPlan. Removed unnecessary code and used functionality from code that already exists in Aerie.

* refactor: use start_time_ms

Update pack method to use start_time_ms since calculation has been fixed, this lowers the amount of computation done in the method itself. Remove null typing for start_time_ms as it theoretically should never be null.

* test: add tests for packActivities

Unit tests for pack Activities as well as generalizing certain variables to be used for multiple

* test: unit tests for pack functions

Create unit tests for pack functions.
Generalize json objects in activities.test.ts so they can be used for multiple unit tests at once.

* refactor:type of start_time_ms for Activity Directives

Change type of start_time_ms attribute for Activity Directives so it just number. Previously, it was number | null but there should be no instance where the start_time_ms is now null.

* refactor: remove await

Pull out repeated actions into a method and update since findTypes is no longer async

* style: vite file

Used npm run format:write to fix code style issues that caused a lint error on the PR run

* refactor: previous tests for new start_time_ms

Refactor old tests to match the new type change for the start_time_ms attribute of Activity Directives.

* RefactorL Migrate PackActivites to ActivityDirectivesTable

Migrate everything related to the packActivitiesDirectives display and modal to ActivityDirectivesTable. Do based off feedback on bulk shifting activity directives PR

* Refactor: packActivities display

Check if there are multiple ActivityDirectives selected then display pack activities options. Change BulkActionDataGrid so new slots are rendered after the predefined ones.

* Refactor: Remove in-place mutations

Change packActivityDirectivesInPlan so it no longer mutates start_offset attributes in place but returns a new list of ActivityDirectives.

* style: documentation

Remove console log statements and unnecessary comments

* Refactor: PR Request Updates

Remove nested function declaration for updateAnchorStartOffset and add appropriate test functions. Change BulkActionDataGrid and SingleActionDataGrid with an additional slot for new functions.

* remove .env from gitignore, remove empty unused credentials from .env

* small improvements to activity packing UI code

* rename context-menu-alt to context-menu-bottom

* add get functions for test fixtures in activities.test.ts instead of globals for safety

* fix linting issues

* move Paste Activities back to top section of context menu

* fix usToOffset function, add unit tests, move padNumber util to text.ts, remove findTypes util in favor of one-liner

* fix activity unit tests for packing, remove unused import

* update getters in ActivityDirectivesTable, add unit test for padNumber

* fix linter issues

* fix linter issue

---------

Co-authored-by: shreyasyellenki <shreyasyellenki@gmail.com>