New Feature: Bulk Shifting Activities (#1721)
* Beginning of PackLeftAction

Working towards creating the packLeft function, and helper methods to parse Stings of format "HH:mm:ss.SSSSSS" into microseconds. There seems to be some issues with the hasura query and the format of the duration string that needs to be reworked.

* Pack Left

Have an early working implementation of packLeft, it currently works on the framework of activities can only be packed if they have the same anchor, "null" is a valid anchor.

* Revamping packLeft

Working on a new implementation of packLeft that allows packing of various anchored activities but follows a specific set of defined rules discussed in a meeting. Will add documentation later to explain the rules

* Reworked Pack Left

Have a new working implementation of packLeft that covers the edge cases of anchors. Some refactoring still needs to be done and then replication for packRight, also realized that initialStartTime is only necessary to keep track of for the first activity and not a map to all activities.

* Reset Pack Activities

Revamped packLeft with the new guidelines involving anchor activities discussed in the Aerie-External Sim Shallow Integration Working Group Meeting. This new approach is in a function with the temporary name packActivityDirectivesInPlanRevamp. I have some ideas to use the same method for left and right with an enum switch between the two but that is something I am still testing, hence the reason it is in a temporary named funciton for now.

* fixed packRight and refactored the appropriate packLeft

Have a working implementation of packRight and updated condition check for if the activity being packed was an anchor to other activities (it initally didn't do anything because it was the wrong conditional). Removed the old packLeftActivityDirectivesInPlan implementation and refactored the "revamp" method to the appropriate name.

* Introduced a Less cluttered version of packing Activities

Combined both implementations of packing Activities in a single method using a const union type for direction. Will be talking with Aerie-UI team if this is fine for the development standards, for now not deleting the code for the separate options as well.

* Started building PackActivitiesOffsetSelect

Built a dialog for packing activity directive with offset, the cancel functionality works but the pack button still needs to be updated. First need to make some refactoring changes to pack actions.

* Code cleanup for packActivityDirectivesInPlan

Deleted the separate methods for packing activities to left and right, and kept packActivityDirectivesInPlan which handles both cases. Removed unnecessary code and used aerie-ui stores instead to achieve same functionality.

* Added offset functionality

Instantiated an offset field (in microseconds) to packing Activities. This will play a more vital role in the development of the pack-actions-offset feature that is also being developed. When the regular 'Pack Left' and 'Pack Right' options are selected the offset field is 0.

* Working Implementation of Pack with Offset

Have a working implementation of packing activities with offset. A dialog pops up that allows the user to pick a direction and offset (default to 0).

* Updated design for PackActivitiesOffsetSelect

Changed the offset string box to the appropriate CSS for error, don't allow negative offsets to be entered. The user cannot pack if the Offset string input is invalid.

* Working Implementation of Bulk Shifting Activities

Have a working implementation of bulk shifting activities. The user receives a dialog allowing them to pick a direction and duration (default to 0) to shift. The user can only shift if they enter a valid duration.

* Changed location of where database gets updated

Felt it would be more appropriate if the database was updated from ActivityDirectivesTable and packActivityDirectivesInPlan just returns the array of updated activities instead. This also simplifies when writing unit tests

* Update Packing with Offset to match new packing framework

Updated packActivityWithOffset so it the updates to the database are done in the ActivityDirectivesTable, also simplified updating an array of ActivityDirectives into one method

* Merged with packActivitiesWithOffset

Merged the two branches and staying with the universal idea that updating the database is done outside of the activities.ts and rather inside ActivityDirectivesTable. Some more changes need to be made for bulk shifting activities to make that possible.

* Fixed directional mishap for packLeft

During refactoring to move updating the database outside of the utility function itself, packLeft was accidentally affected so it would move right. That has now been fixed.

* refactor: modal design

Change the current display modal for bulk shift to utilize the existing Aerie modal in svelte

* refactor: bulk modal design

Change design aspects of bulk shift modal, will get it reviewed by Aerie team members

* refactor: bulk shift workflow

Change bulkShiftActivityDirectivesInPlan so it only returns the updated activities. The Database update is now in ActivityDirectiveTable.svelte to match the packing functions. Minor design changes to BulkdShiftActivitiesModal, now has a header and buttton and input height are the same.

* refactor: remove async, test: add bulkShiftActivityDirectivesInPlan

Since, the database update happens outside bulkShiftActivityDirectivesInPlan it no longer has to be an async function. Added unit for bulk shift for both left and right direction.

* Refactor: remove pack Activity Directives code

Delete all code related to pack Activities to make code review easier

* refactor: fix change made to test

Revert accidental change made to a addAbsoluteTimeToRevision unit test

* Style: Update for PR

Change bulkShiftActivityDirectivesInPlan so it is no longer in place. Make style changes to the modal as well as unit tests for new utility functions.

* Refactor: Move BulkShift Modal display

Migrate the display of the bulk shift modal into ActivityDirectivesTable rather than BulkActionDataGrid.

* Refactor: Extra slot in BulkActionDataGrid

Move bulk shift option display into extra slot of BulkActionDataGrid. Did this to match PR#1720

* revert changes to .env and remove from gitignore

* run linter, small changes from review, remove unused findTypes util

* fix context menu slots

* move 'pack activities' logic and modal creation to effects

* fix linter

* fix linter

* refactor shift activities to effects and use consistent modal pattern

* await results of async pack/shift activities

* standardize button styles in pack/shift modals

---------

Co-authored-by: dandelany <dan.delany@gmail.com>